fix cleanup and some bugs when there are 0 articles

+++ cleanup.py
@@ -6,8 +6,16 @@ for image in images:
     os.remove(image)
 
 #remove db
-os.remove("wiki_articles.db")
+#check if file exists
+if os.path.exists("wiki_articles.db"):
+    os.remove("wiki_articles.db")
 
 #remove the directory chromadb/index and its contents
 import shutil
-shutil.rmtree("chromadb/index")
+#check if directory exists
+if os.path.exists("chromadb/index"):
+    shutil.rmtree("chromadb/index")
+if os.path.exists("chromadb/chroma-collections.parquet"):
+    os.remove("chromadb/chroma-collections.parquet")
+if os.path.exists("chromadb/chroma-embeddings.parquet"):
+    os.remove("chromadb/chroma-embeddings.parquet")
